Code refactor for source code parser
Renamed sourced code parser to ObjectMiner.
This commit improve performance and fixes bugs in ObjectMiner tool.
All code was divided to smaller pieces, added interfaces and extensions.
Added converters, that converting parsed data to JSON or XML file.
Added sample converter, that displays data in console as a list, and sample extensions, that brings possibility to display data in console and sleep.